+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # rlang 0.2.0.9000
 
+* Functions with tidy dots support now allow splicing atomic vectors.
+
 * Formulas are now evaluated in the correct environment within
   `eval_tidy()`. This fixes issues in dplyr and other tidy-evaluation
   interfaces.